cloud-functions/insert-airline-review/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.25.0
flask==3.0.3
